calculator
There is a class named Calculator. It has a member function getResult() which can calculate the value of the expression. Please complete the design and implementation of the class.
Something you should be noticed:
(1)The expression includes only digits and symbol +, -, *, /, (, and ).
(2)The expression is guaranteed to be valid.
 
Notice: The operation / is like the division in our code. For example, 3/4=0 and 6/5=1.